move slack oauth logic into slack integration
