--- busy_beaver/apps/oauth_integrations/oauth_providers/__init__.py
@@ -1,1 +1,0 @@
-from .slack import StateToOAuthResponse  # noqa
